requirements.txt
ml-collections==0.1.1
absl-py==1.3.0
datasets==2.6.2
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
pyarrow==9.0.0
pytorch-lightning==1.8.6
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
wandb==0.13.5

